backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
mysql-connector-python
scikit-learn
pandas
python-dotenv
joblib
gunicorn
jinja2
authlib>=1.2.0
httpx>=0.27.0
aiokafka>=0.10.0
elasticsearch[async]>=8.13.0
mlflow>=2.15.1
aiosmtplib>=3.0.0
email-validator>=2.0.0
numpy

